requirements.txt

fastapi
uvicorn[standard]
sqlmodel
asyncpg
python-jose[cryptography]
passlib[bcrypt]
firebase-admin
python-dotenv
